requirements.txt
########## LLM Framework ##########
# https://docs.llamaindex.ai/en/stable
llama-index==0.10.55
########## Database ##########
chromadb==0.5.3
# https://llamahub.ai/l/vector_stores/llama-index-vector-stores-chroma?from=
# llama-index-vector-stores-chroma==0.1.10
# https://llamahub.ai/l/graph_stores/llama-index-graph-stores-neo4j?from=
llama-index-graph-stores-neo4j==0.2.6
########## Other ##########
typing-extensions==4.9.0
python-dotenv==1.0.1
requests~=2.32.3
pydantic~=2.8.2
neo4j~=5.22.0
mysqlclient==2.2.4
SQLAlchemy==2.0.31
fastapi==0.112.1
uvicorn==0.30.1
starlette==0.37.2